data-pipeline/ingest.py
import os
from pathlib import Path
from dotenv import load_dotenv

load_dotenv()

class DataPipeline:
    def __init__(self):
        self.raw_docs_dir = Path("raw_docs")
        self.processed_dir = Path("processed")
    
    def parse_pdfs(self):
        # Your LlamaParse logic here
        pass
    
    def chunk_documents(self):
        # Your chunking logic here
        pass
    
    def generate_embeddings(self):
        # Your embedding logic here
        pass
    
    def upload_to_supabase(self):
        # Your Supabase upload logic here
        pass

if __name__ == "__main__":
    pipeline = DataPipeline()
    # Run your pipeline steps here
